example/gemini/requirements.txt
aiobs
google-genai>=1.0.0
google-auth>=2.0.0

